set distribute test command and 2to3 option
To be converted to python 3, the tests and example needed to be
added to the setup package option. So I added them to the 
pystache namespace.

I then had to correct the relative module path and template path.
The template path need to be absolute but it might not be
necessary.